in progress
still need to finish migrating functionality, need to adapt to new initializeModel functionality (i.e. storyteller needs to be moved back to dixit so that it can access model variables for generation. Technically can change it back so we just have a player class and have a boolean bot or not variable.